requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-multipart>=0.0.6
requests>=2.18.4